# services/control-api/openapi.yaml
openapi: 3.1.0
info:
  title: Control API
  version: 0.1.0
  description: >
    Control-plane API for managing workspaces, pipelines, connectors, schemas, and observability.
    Provides resource CRUD and Kafka admin operations for Event Integration Platform.
  contact:
    name: Platform Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://localhost:8080/api
    description: API base path

security:
  - BearerAuth: []

paths:
  /health:
    get:
      summary: Service health check
      operationId: getHealth
      security: []
      tags:
        - Health
      responses:
        "200":
          description: Service is healthy
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/HealthResponse

  /api/health:
    get:
      summary: API health check
      operationId: getApiHealth
      security: []
      tags:
        - Health
      responses:
        "200":
          $ref: "#/paths/~1health/get/responses/200"

  /api/workspaces:
    get:
      summary: List all workspaces
      operationId: listWorkspaces
      tags:
        - Workspaces
      parameters:
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PageQuery
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PerPageQuery
      responses:
        "200":
          description: List of workspaces
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/ListResponse
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/Workspace
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

    post:
      summary: Create a new workspace
      operationId: createWorkspace
      tags:
        - Workspaces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/CreateWorkspaceRequest
      responses:
        "201":
          description: Workspace created successfully
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/Workspace
        "400":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/ValidationError
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

  /api/pipelines:
    get:
      summary: List all pipelines
      operationId: listPipelines
      tags:
        - Pipelines
      parameters:
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PageQuery
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PerPageQuery
      responses:
        "200":
          description: List of pipelines
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/ListResponse
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/Pipeline
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

    post:
      summary: Create a new pipeline
      operationId: createPipeline
      tags:
        - Pipelines
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/CreatePipelineRequest
      responses:
        "201":
          description: Pipeline created successfully
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/Pipeline
        "400":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/ValidationError
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

  /api/pipelines/{pipelineId}:
    get:
      summary: Get a single pipeline by ID
      operationId: getPipeline
      tags:
        - Pipelines
      parameters:
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PipelineIdParam
      responses:
        "200":
          description: Pipeline details
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/Pipeline
        "400":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/BadRequest
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "404":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/NotFound
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

    put:
      summary: Update a pipeline
      operationId: updatePipeline
      tags:
        - Pipelines
      parameters:
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PipelineIdParam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/UpdatePipelineRequest
      responses:
        "200":
          description: Pipeline updated successfully
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/Pipeline
        "400":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/ValidationError
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "404":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/NotFound
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

  /api/workspaces/{workspaceId}/clients:
    get:
      summary: List clients in a workspace
      operationId: listClients
      tags:
        - Clients
      parameters:
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/WorkspaceIdParam
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PageQuery
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PerPageQuery
      responses:
        "200":
          description: List of clients
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/ListResponse
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/Client
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "404":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/NotFound
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

    post:
      summary: Create a new client in a workspace
      operationId: createClient
      tags:
        - Clients
      parameters:
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/WorkspaceIdParam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/CreateClientRequest
      responses:
        "201":
          description: Client created successfully
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/Client
        "400":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/ValidationError
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "404":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/NotFound
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

  /api/workspaces/{workspaceId}/users:
    get:
      summary: List users in a workspace
      operationId: listUsers
      tags:
        - Users
      parameters:
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/WorkspaceIdParam
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PageQuery
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PerPageQuery
      responses:
        "200":
          description: List of users
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/ListResponse
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/User
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "404":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/NotFound
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

    post:
      summary: Create a new user in a workspace
      operationId: createUser
      tags:
        - Users
      parameters:
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/WorkspaceIdParam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/CreateUserRequest
      responses:
        "201":
          description: User created successfully
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/User
        "400":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/ValidationError
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "404":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/NotFound
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

  /api/topics:
    get:
      summary: List Kafka topics
      operationId: listTopics
      tags:
        - Topics
      responses:
        "200":
          description: List of topics
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/TopicName
                        partitions:
                          type: integer
                        replicationFactor:
                          type: integer
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

    post:
      summary: Create a Kafka topic
      operationId: createTopic
      tags:
        - Topics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/CreateTopicRequest
      responses:
        "201":
          description: Topic created successfully
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/TopicName
                  partitions:
                    type: integer
                  replicationFactor:
                    type: integer
        "400":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/ValidationError
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "409":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Conflict
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

  /api/topics/{topicName}/metrics:
    get:
      summary: Get metrics for a specific topic
      operationId: getTopicMetrics
      tags:
        - Topics
      parameters:
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/TopicNameParam
      responses:
        "200":
          description: Topic metrics
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/TopicMetrics
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "404":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/NotFound
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

  /api/workspaces/{workspaceId}/jsonata-transforms:
    get:
      summary: List Jsonata transforms in a workspace
      operationId: listJsonataTransforms
      tags:
        - Transforms
      parameters:
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/WorkspaceIdParam
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PageQuery
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/PerPageQuery
      responses:
        "200":
          description: List of Jsonata transforms
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/ListResponse
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/JsonataTransform
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "404":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/NotFound
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

    post:
      summary: Create a new Jsonata transform
      operationId: createJsonataTransform
      tags:
        - Transforms
      parameters:
        - $ref: ../../packages/openapi-components/control-api.yaml#/components/parameters/WorkspaceIdParam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/CreateJsonataTransformRequest
      responses:
        "201":
          description: Transform created successfully
          headers:
            x-request-id:
              $ref: ../../packages/openapi-components/control-api.yaml#/components/headers/RequestId
          content:
            application/json:
              schema:
                $ref: ../../packages/openapi-components/control-api.yaml#/components/schemas/JsonataTransform
        "400":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/ValidationError
        "401":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/Unauthorized
        "404":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/NotFound
        "500":
          $ref: ../../packages/openapi-components/control-api.yaml#/components/responses/InternalServerError

components:
  securitySchemes:
    BearerAuth:
      $ref: ../../packages/openapi-components/control-api.yaml#/components/securitySchemes/BearerAuth

tags:
  - name: Health
    description: Service health endpoints
  - name: Workspaces
    description: Workspace management
  - name: Pipelines
    description: Pipeline lifecycle and configuration
  - name: Clients
    description: Client (machine identity) management
  - name: Users
    description: User management
  - name: Topics
    description: Kafka topic administration
  - name: Transforms
    description: Jsonata transformation configuration